% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-gh.R
\name{fetch_all_comments}
\alias{fetch_all_comments}
\title{Fetch all comments for an issue or PR}
\usage{
fetch_all_comments(owner, repo, number, resource_type)
}
\arguments{
\item{owner}{Repository owner}

\item{repo}{Repository name}

\item{number}{Issue or PR number}

\item{resource_type}{Either "issue" or "pr"}
}
\value{
List of comments
}
\description{
Fetch all comments for an issue or PR
}
\keyword{internal}
