% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-gh.R
\name{fetch_all_review_comments}
\alias{fetch_all_review_comments}
\title{Fetch all review comments for a PR}
\usage{
fetch_all_review_comments(owner, repo, number)
}
\arguments{
\item{owner}{Repository owner}

\item{repo}{Repository name}

\item{number}{PR number}
}
\value{
List of review comments
}
\description{
Fetch all review comments for a PR
}
\keyword{internal}
