% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-gh.R
\name{fetch_issue}
\alias{fetch_issue}
\title{Fetch issue data using gh package}
\usage{
fetch_issue(owner, repo, number)
}
\arguments{
\item{owner}{Repository owner}

\item{repo}{Repository name}

\item{number}{Issue number}
}
\value{
Issue data as list
}
\description{
Fetch issue data using gh package
}
\keyword{internal}
