% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-gh.R
\name{fetch_pull_request}
\alias{fetch_pull_request}
\title{Fetch pull request data using gh package}
\usage{
fetch_pull_request(owner, repo, number)
}
\arguments{
\item{owner}{Repository owner}

\item{repo}{Repository name}

\item{number}{PR number}
}
\value{
PR data as list
}
\description{
Fetch pull request data using gh package
}
\keyword{internal}
