% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stories.R
\docType{package}
\name{stories}
\alias{stories-package}
\alias{stories}
\title{Make stories from your histories}
\description{
Uses AI to make sense of git histories.
}
\author{
\strong{Maintainer}: Lionel Henry \email{lionel@posit.co}

}
